+++ modules/portal/conf/application-test.conf
@@ -7,21 +7,21 @@ portal.vinyldns.backend.url = "http://not.real.com"
 dynamo {
   key = "akid goes here"
   secret = "secret key goes here"
-  endpoint = "http://127.0.0.1:19000"
+  endpoint = "http://foo.bar"
   region = "us-east-1" # note: we are always in us-east-1, but this can be overridden
   test_datastore = false
 }
 
 users {
   dummy = false
-  tablename = "users"
+  tablename = "users-test"
   provisionedReadThroughput = 100
   provisionedWriteThroughput = 100
 }
 
 changelog {
   dummy = false
-  tablename = "usersAndGroupChanges"
+  tablename = "usersAndGroupChanges-test"
   provisionedReadThroughput = 100
   provisionedWriteThroughput = 100
 }
@@ -60,3 +60,5 @@ links = [
     icon = ""
   }
 ]
+
+play.modules.enabled += "modules.VinylDNSModule"
