+++ modules/portal/conf/application.conf
@@ -111,5 +111,7 @@ links = [
   }
 ]
 
+play.modules.enabled += "modules.VinylDNSModule"
+
 // Local.conf has files specific to your environment, for example your own LDAP settings
 include "local.conf"
